+++ src/time/mktime.c
@@ -27,7 +27,16 @@ time_t mktime(struct tm *tm)
 	if (mon < 0) { mon += 12; y--; }
 	m = (unsigned)mon + 1;
 
-	days = __days_from_civil(y, m, 1) + (tm->tm_mday - 1);
+	/* tm_mday gets the same widening as tm_hour/tm_min below, and for
+	 * the same reason: it is an `int` the caller is entitled to hand
+	 * over out of range, and `tm->tm_mday - 1` is evaluated in `int`.
+	 * At tm_mday == INT_MIN that subtraction is signed overflow --
+	 * undefined behaviour inside the very function whose contract is
+	 * that out-of-range fields are accepted and normalized.  Found by
+	 * fuzz/fuzz_time.c; UBSan reported
+	 * "signed integer overflow: -2147483648 - 1 cannot be represented
+	 * in type 'int'" at this line. */
+	days = __days_from_civil(y, m, 1) + ((long long)tm->tm_mday - 1);
 	/* tm_hour/tm_min are `int`, unbounded like tm_mon above (mktime must
 	 * accept and normalize out-of-range fields); widen before multiplying
 	 * so a large caller-supplied value overflows in `long long`, not in
